feat: refactor value converters and implement ExtractValues for Row
